requirements.txt
openai>=1.3.0
chromadb>=0.4.18
python-dotenv>=1.0.0
tiktoken>=0.5.2
PyPDF2>=3.0.1
python-docx>=0.8.11